+++ src/main/java/app/repository/mapper/SongMapper.java
@@ -4,6 +4,7 @@ import app.model.Song;
 import org.jdbi.v3.core.mapper.RowMapper;
 import org.jdbi.v3.core.statement.StatementContext;
 
+import java.math.BigDecimal;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
@@ -16,7 +17,7 @@ public class SongMapper implements RowMapper<Song> {
         String artist = r.getString("SONG.artist");
         Integer streams = r.getInt("SONG.streams");
         String url = r.getString("SONG.url");
-        Integer price = r.getInt("SONG.price");
+        BigDecimal price = r.getBigDecimal("SONG.price");
         String date = r.getString("SONG.date");
 
         return new Song(id, position, trackName, artist, streams, url, price, date);
