+++ src/main/java/app/repository/mapper/UserMapper.java
@@ -4,6 +4,7 @@ import app.model.User;
 import org.jdbi.v3.core.mapper.RowMapper;
 import org.jdbi.v3.core.statement.StatementContext;
 
+import java.math.BigDecimal;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
@@ -12,7 +13,7 @@ public class UserMapper implements RowMapper<User> {
     public User map(ResultSet r, StatementContext ctx) throws SQLException {
         String id = r.getString("USER.id");
         String username = r.getString("USER.username");
-        Double balance = r.getDouble("USER.balance");
+        BigDecimal balance = r.getBigDecimal("USER.balance");
 
 
         return new User(id, username, balance);
